many new unit tests // insect module computes reynolds numbers+wing area // dry run inits kinematics.t

+++ tests/unittest.sh
@@ -8,13 +8,41 @@ echo "***************************************************"
 #tar xzf tests_data.tar.gz
 
 # list all the test scrits you want, separated by spaces
-tests=(solid_model.sh solid_model2.sh solid_model3.sh solid_model4.sh jerry.sh striding.sh
-       fruitfly_mask.sh fruitfly_mask_movingbody.sh fruitfly_mask_kineloader.sh bumblebee_mask.sh scalar.sh
-       vortexring1_AB2.sh
-       vortexring2_RK2.sh vortexring4_sponge.sh
-       sphere_sponge_RK2.sh sphere.sh sphere_restart.sh sphere_fall_fsi.sh mhdorszagtang.sh
-       insect.sh insect_RK4.sh 
-       swimmer1_iteration.sh swimmer2_staggered.sh swimmer3_semiimplicit.sh)
+tests=( '---insect-module---'
+insects_mask_FF_tethered.sh
+insects_mask_BB_tethered.sh
+insects_mask_FF_tethered_bodyrotated.sh
+insects_mask_FF_tethered_kineloader.sh
+insects_mask_FF_moving_rotation.sh
+insects_mask_FF_moving_rotationtranslation.sh
+insects_mask_jerry.sh
+insects_flow_FF_tethered.sh
+insects_flow_BB_tethered_AB2.sh
+insects_flow_BB_tethered_AB2_periodic.sh
+insects_flow_BB_tethered_RK4.sh
+insects_flow_COIN_freeflight.sh
+insects_flow_FF_freeflight.sh
+'---fluid-without-penalization---'
+vortexring1_AB2.sh
+vortexring2_RK2.sh
+vortexring4_sponge.sh
+'---fluid-with-simple-penalization---'
+sphere.sh
+sphere_restart.sh
+sphere_sponge_RK2.sh
+'---fluid-structure-interaction---'
+solid_model.sh
+solid_model2.sh
+solid_model3.sh
+solid_model4.sh
+swimmer1_iteration.sh
+swimmer2_staggered.sh
+swimmer3_semiimplicit.sh
+'---Misc---'
+mhdorszagtang.sh
+striding.sh
+scalar.sh
+)
 
 # link flusi and mhd from .. to . if this isn't already done.
 if [ ! -f flusi ]; then
@@ -49,30 +77,33 @@ for test in ${tests[*]}
 do
     numtests=$(($numtests + 1))
 
-    logfile=${test%%.sh}.log
-    rm -f $logfile
-    touch $logfile
+    if [[ $test == "---"* ]]; then
+    	echo ""
+    	echo $test
+    	echo ""
+  	else
+        logfile=${test%%.sh}.log
+        rm -f $logfile
+        touch $logfile
 
-    # run the test, copy the output to the logfile
-    ./$test > $logfile
-    success=$?
-    if [ $success == 0 ]
-    then
-	echo -e "OK\tRunning test: "${test}", log in: "${logfile}
-	numsuccess=$(($numsuccess + 1))
-    else
-	echo -e "FAIL\tRunning test: "${test}", log in: "${logfile}
-	numfail=$(($numfail + 1))
+        # make sure no old data is present before running test
+        rm -f *.h5 *.t *.ini
+
+        # run the test, copy the output to the logfile
+        ./$test > $logfile
+        success=$?
+        if [ $success == 0 ]; then
+    	     echo -e "OK\tRunning test: "${test}", log in: "${logfile}
+    	     numsuccess=$(($numsuccess + 1))
+        else
+    	     echo -e "FAIL\tRunning test: "${test}", log in: "${logfile}
+    	     numfail=$(($numfail + 1))
+        fi
+
+        #cleanup
+        rm -f *.key *.h5 *.t succ* decom*
+        rm -f runtime*
     fi
-
-    #cleanup
-    rm -f *.key
-    rm -f *.h5
-    rm -f drag_data
-    rm -f *.t end IBES_iter
-    rm -f runtime*.ini
-    rm -f success
-    rm -f deco*
 done
 
 echo
@@ -80,7 +111,7 @@ T="$(($(date +%s)-T))"
 echo "Time used in tests: ${T} seconds"
 
 echo
-echo "Total number of tests: " $numtests
+echo "Total number of tests: " $(($numsuccess+$numfail))
 echo "Total number successful: " $numsuccess
 echo "Total number failed: " $numfail
